Extract deploy/app.go
