+++ internal/pkg/deploy/deploy.go
@@ -4,19 +4,6 @@
 // Package deploy holds the structures to deploy infrastructure resources.
 package deploy
 
-import (
-	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/archer"
-	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/manifest"
-)
-
-// CreateLBFargateAppInput holds the fields required to deploy a load-balanced AWS Fargate application.
-type CreateLBFargateAppInput struct {
-	App          *manifest.LBFargateManifest
-	Env          *archer.Environment
-	ImageRepoURL string
-	ImageTag     string
-}
-
 // Resource represents an AWS resource.
 type Resource struct {
 	LogicalName string
